refactor: rename "discovery" to "plugins"
